Issue #20747: move rule43onestatement inputs to compilable folder
