+++ config/google-java-format/excluded/noncompilable-input-paths.txt
@@ -1,1 +1,1 @@
-src/it/resources-noncompilable/com/google/checkstyle/test/chapter4formatting/rule43onestatement/InputOneStatementPerLine2.java
+
