+++ src/it/resources/com/sun/checkstyle/test/chapter7statements/rule71simplestatements/InputOneStatementPerLineDefault.java
@@ -1,7 +1,4 @@
-package com.sun.checkstyle.test.chapter7statements.rule71simplestatements;
-
-// violation first line 'Header mismatch'
-
+package com.sun.checkstyle.test.chapter7statements.rule71simplestatements;// violation first line 'Header mismatch'
 /**
  * Test input for one statement per line with default violations.
  */
