+++ src/it/resources/com/sun/checkstyle/test/chapter7statements/rule71simplestatements/InputOneStatementPerLineValid.java
@@ -1,7 +1,4 @@
-package com.sun.checkstyle.test.chapter7statements.rule71simplestatements;
-
-// violation first line 'Header mismatch'
-
+package com.sun.checkstyle.test.chapter7statements.rule71simplestatements;// violation first line 'Header mismatch'
 /**
  * Test input for valid one statement per line.
  */
